fix(app): Combining the EVM and OAuth sidebar components into one
