--- app/components/GenericOauthPlatform.tsx
@@ -1,218 +1,0 @@
-// --- Methods
-import React, { useContext, useEffect, useMemo, useState } from "react";
-
-// --- Datadog
-import { datadogLogs } from "@datadog/browser-logs";
-
-import { debounce } from "ts-debounce";
-import { BroadcastChannel } from "broadcast-channel";
-
-// --- Identity tools
-import {
-  Stamp,
-  VerifiableCredential,
-  CredentialResponseBody,
-  VerifiableCredentialRecord,
-} from "@gitcoin/passport-types";
-import { fetchVerifiableCredential } from "@gitcoin/passport-identity/dist/commonjs/src/credentials";
-
-// --- Style Components
-import { SideBarContent } from "./SideBarContent";
-import { DoneToastContent } from "./DoneToastContent";
-import { useToast } from "@chakra-ui/react";
-
-// --- Context
-import { CeramicContext } from "../context/ceramicContext";
-import { UserContext } from "../context/userContext";
-
-// --- Types
-import { PlatformGroupSpec } from "@gitcoin/passport-platforms/dist/commonjs/src/types";
-import {
-  Platform,
-  AccessTokenResult,
-  Proofs,
-  AppContext,
-  ProviderPayload,
-} from "@gitcoin/passport-platforms/dist/commonjs/src/types";
-import { getPlatformSpec, PROVIDER_ID } from "@gitcoin/passport-platforms/dist/commonjs/src/platforms-config";
-
-export type PlatformProps = {
-  // platformId: string;
-  platformgroupspec: PlatformGroupSpec[];
-  platform: Platform;
-};
-
-function generateUID(length: number) {
-  return window
-    .btoa(
-      Array.from(window.crypto.getRandomValues(new Uint8Array(length * 2)))
-        .map((b) => String.fromCharCode(b))
-        .join("")
-    )
-    .replace(/[+/]/g, "")
-    .substring(0, length);
-}
-
-export const GenericOauthPlatform = ({ platformgroupspec, platform }: PlatformProps): JSX.Element => {
-  const { address, signer } = useContext(UserContext);
-  const { handleAddStamps, allProvidersState } = useContext(CeramicContext);
-  const [isLoading, setLoading] = useState(false);
-  const [canSubmit, setCanSubmit] = useState(false);
-
-  // find all providerIds
-  const providerIds = useMemo(
-    () =>
-      platformgroupspec?.reduce((all, stamp) => {
-        return all.concat(stamp.providers?.map((provider) => provider.name as PROVIDER_ID));
-      }, [] as PROVIDER_ID[]) || [],
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  );
-
-  // SelectedProviders will be passed in to the sidebar to be filled there...
-  const [verifiedProviders, setVerifiedProviders] = useState<PROVIDER_ID[]>(
-    providerIds.filter((providerId) => {
-      return typeof allProvidersState[providerId]?.stamp?.credential !== "undefined";
-    })
-  );
-  // SelectedProviders will be passed in to the sidebar to be filled there...
-  const [selectedProviders, setSelectedProviders] = useState<PROVIDER_ID[]>([...verifiedProviders]);
-
-  // any time we change selection state...
-  useEffect(() => {
-    if (selectedProviders.length !== verifiedProviders.length) {
-      setCanSubmit(true);
-    }
-    if (selectedProviders.length === 0) {
-      setCanSubmit(false);
-    }
-  }, [selectedProviders, verifiedProviders]);
-
-  // --- Chakra functions
-  const toast = useToast();
-
-  async function fetchCredential(proofs: Proofs): Promise<void> {
-    // fetch VCs for only the selectedProviders
-    const vcs = await fetchVerifiableCredential(
-      process.env.NEXT_PUBLIC_PASSPORT_IAM_URL || "",
-      {
-        type: platform.platformId,
-        types: selectedProviders,
-        version: "0.0.0",
-        address: address || "",
-        proofs,
-      },
-      signer as { signMessage: (message: string) => Promise<string> }
-    );
-    await setVerifiedCredentialState(vcs);
-  }
-
-  async function setVerifiedCredentialState(verified: VerifiableCredentialRecord) {
-    // because we provided a types array in the params we expect to receive a credentials array in the response...
-    const vcs =
-      verified.credentials
-        ?.map((cred: CredentialResponseBody): Stamp | undefined => {
-          if (!cred.error) {
-            // add each of the requested/received stamps to the passport...
-            return {
-              provider: cred.record?.type as PROVIDER_ID,
-              credential: cred.credential as VerifiableCredential,
-            };
-          }
-        })
-        .filter((v: Stamp | undefined) => v) || [];
-    // Add all the stamps to the passport at once
-    await handleAddStamps(vcs as Stamp[]);
-    // report success to datadog
-    datadogLogs.logger.info("Successfully saved Stamp", { platform: platform.platformId });
-    // grab all providers who are verified from the verify response
-    const actualVerifiedProviders = providerIds.filter(
-      (providerId: string | undefined) =>
-        !!vcs.find((vc: Stamp | undefined) => vc?.credential?.credentialSubject?.provider === providerId)
-    );
-    // both verified and selected should look the same after save
-    setVerifiedProviders([...actualVerifiedProviders]);
-    setSelectedProviders([...actualVerifiedProviders]);
-    // reset can submit state
-    setCanSubmit(false);
-
-    toast({
-      duration: 5000,
-      isClosable: true,
-      render: (result: any) => (
-        <DoneToastContent
-          title="Success!"
-          body={`All ${platform.platformId} data points verified.`}
-          icon="../../assets/check-icon.svg"
-          platformId={platform.platformId as PLATFORM_ID}
-          result={result}
-        />
-      ),
-    });
-  }
-
-  const state = `${platform.path}-` + generateUID(10);
-
-  const waitForRedirect = (timeout?: number): Promise<ProviderPayload> => {
-    const channel = new BroadcastChannel(`${platform.path}_oauth_channel`);
-    const waitForRedirect = new Promise<ProviderPayload>((resolve, reject) => {
-      // Listener to watch for oauth redirect response on other windows (on the same host)
-      function listenForRedirect(e: { target: string; data: { code: string; state: string } }) {
-        // when receiving oauth response from a spawned child run fetchVerifiableCredential
-        if (e.target === platform.path) {
-          // pull data from message
-          const queryCode = e.data.code;
-          const queryState = e.data.state;
-          datadogLogs.logger.info("Saving Stamp", { platform: platform.platformId });
-          try {
-            resolve({ code: queryCode, state: queryState });
-          } catch (e) {
-            datadogLogs.logger.error("Error saving Stamp", { platform: platform.platformId });
-            console.error(e);
-            reject(e);
-          }
-        }
-      }
-      // event handler will listen for messages from the child (debounced to avoid multiple submissions)
-      channel.onmessage = debounce(listenForRedirect, 300);
-    }).finally(() => {
-      channel.close();
-    });
-    return waitForRedirect;
-  };
-
-  async function initiateFetchCredential() {
-    try {
-      // fetch and store credential
-      setLoading(true);
-      const providerPayload = await platform.getProviderPayload({ state, window, screen, waitForRedirect });
-      // TODO: use only one of Proofs or ProfiderPayload and drop the other
-      await fetchCredential(providerPayload as unknown as Proofs);
-    } catch (e) {
-      datadogLogs.logger.error("Error saving Stamp", { platform: platform.platformId });
-      console.error(e);
-    }
-    setLoading(false);
-  }
-
-  return (
-    <SideBarContent
-      currentPlatform={getPlatformSpec(platform.platformId)}
-      currentProviders={platformgroupspec}
-      verifiedProviders={verifiedProviders}
-      selectedProviders={selectedProviders}
-      setSelectedProviders={setSelectedProviders}
-      isLoading={isLoading}
-      verifyButton={
-        <button
-          disabled={!canSubmit}
-          onClick={initiateFetchCredential}
-          data-testid={`button-verify-${platform.platformId}`}
-          className="sidebar-verify-btn"
-        >
-          Verify
-        </button>
-      }
-    />
-  );
-};
