+++ platforms/src/Gitcoin/App-Bindings.ts
@@ -7,19 +7,37 @@ export class GitcoinPlatform implements Platform {
   clientId: string = null;
   redirectUri: string = null;
 
-  async getProviderPayload(appContext: AppContext): Promise<ProviderPayload> {
-    return {};
-  }
-
-  
   constructor(options: PlatformOptions = {}) {
     this.clientId = options.clientId as string;
     this.redirectUri = options.redirectUri as string;
   }
 
+  async getProviderPayload(appContext: AppContext): Promise<ProviderPayload> {
+    const authUrl: string = await this.getOAuthUrl(appContext.state);
+    const width = 600;
+    const height = 800;
+    const left = appContext.screen.width / 2 - width / 2;
+    const top = appContext.screen.height / 2 - height / 2;
+
+    // Pass data to the page via props
+    appContext.window.open(
+      authUrl,
+      "_blank",
+      "toolbar=no, location=no, directories=no, status=no, menubar=no, resizable=no, copyhistory=no, width=" +
+        width +
+        ", height=" +
+        height +
+        ", top=" +
+        top +
+        ", left=" +
+        left
+    );
+
+    return appContext.waitForRedirect();
+  }
+
   async getOAuthUrl(state: string): Promise<string> {
     const githubUrl = `https://github.com/login/oauth/authorize?client_id=${this.clientId}&redirect_uri=${this.redirectUri}&state=${state}`;
-    
     return githubUrl;
   }
 }
